refactor: invalidate hook 추가 및 invalidate 로직 뮤테이션에서 사용처로 분리
