+++ src/hooks/applicants/useInvalidateApplicants.ts
@@ -1,0 +1,7 @@
+import { useQueryClient } from '@tanstack/react-query';
+
+export const useInvalidateApplicants = () => {
+  const queryClient = useQueryClient();
+
+  return () => queryClient.invalidateQueries({ queryKey: ['applicants'] });
+};
