+++ src/hooks/applicants/usePatchApplicant.ts
@@ -1,4 +1,4 @@
-import { useMutation, useQueryClient } from '@tanstack/react-query';
+import { useMutation } from '@tanstack/react-query';
 import { api } from '@/apis/api.ts';
 import { PatchApplicant, PatchApplicantSchema } from '@/scheme/applicant.ts';
 
@@ -8,16 +8,11 @@ interface PatchApplicantParams {
 }
 
 export const usePatchApplicant = () => {
-  const queryClient = useQueryClient();
-
   return useMutation({
     mutationFn: async ({ applicantId, params }: PatchApplicantParams) => {
       await api.patch(`applicants/${applicantId}`, {
         body: JSON.stringify(PatchApplicantSchema.parse(params)),
       });
-    },
-    onSuccess: async () => {
-      await queryClient.invalidateQueries({ queryKey: ['applicants'] });
     },
   });
 };
